+++ stratum/hal/lib/yang/yang_parse_tree_interface.cc
@@ -3,7 +3,7 @@
 // Copyright 2022 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#include "stratum/hal/lib/common/yang_parse_tree_paths.h"
+#include "stratum/hal/lib/yang/yang_parse_tree_paths.h"
 
 #include "absl/container/flat_hash_map.h"
 #include "gnmi/gnmi.pb.h"
@@ -12,8 +12,8 @@
 #include "stratum/hal/lib/common/common.pb.h"
 #include "stratum/hal/lib/common/gnmi_events.h"
 #include "stratum/hal/lib/common/gnmi_publisher.h"
-#include "stratum/hal/lib/common/yang_parse_tree_helpers.h"
-#include "stratum/hal/lib/common/yang_parse_tree.h"
+#include "stratum/hal/lib/yang/yang_parse_tree_helpers.h"
+#include "stratum/hal/lib/yang/yang_parse_tree.h"
 #include "stratum/hal/lib/common/utils.h"
 
 namespace stratum {
