+++ stratum/hal/lib/yang/yang_parse_tree_node.cc
@@ -3,15 +3,15 @@
 // Copyright 2022 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#include "stratum/hal/lib/common/yang_parse_tree_paths.h"
+#include "stratum/hal/lib/yang/yang_parse_tree_paths.h"
 
 #include "absl/strings/str_format.h"
 #include "gnmi/gnmi.pb.h"
 #include "stratum/hal/lib/common/gnmi_events.h"
 #include "stratum/hal/lib/common/gnmi_publisher.h"
-#include "stratum/hal/lib/common/yang_parse_tree_component.h"
-#include "stratum/hal/lib/common/yang_parse_tree_helpers.h"
-#include "stratum/hal/lib/common/yang_parse_tree.h"
+#include "stratum/hal/lib/yang/yang_parse_tree_component.h"
+#include "stratum/hal/lib/yang/yang_parse_tree_helpers.h"
+#include "stratum/hal/lib/yang/yang_parse_tree.h"
 
 namespace stratum {
 namespace hal {
