+++ stratum/hal/lib/yang/yang_parse_tree_paths.cc
@@ -3,7 +3,7 @@
 // Copyright 2022 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-#include "stratum/hal/lib/common/yang_parse_tree_paths.h"
+#include "stratum/hal/lib/yang/yang_parse_tree_paths.h"
 
 #include <utility>
 #include <vector>
@@ -15,8 +15,8 @@
 #include "stratum/hal/lib/common/gnmi_publisher.h"
 #include "stratum/hal/lib/common/openconfig_converter.h"
 #include "stratum/hal/lib/common/utils.h"
-#include "stratum/hal/lib/common/yang_parse_tree_component.h"
-#include "stratum/hal/lib/common/yang_parse_tree_helpers.h"
+#include "stratum/hal/lib/yang/yang_parse_tree_component.h"
+#include "stratum/hal/lib/yang/yang_parse_tree_helpers.h"
 
 namespace stratum {
 namespace hal {
