+++ stratum/hal/lib/yang/yang_parse_tree_paths.h
@@ -8,7 +8,7 @@
 #include <string>
 
 #include "stratum/hal/lib/common/utils.h"
-#include "stratum/hal/lib/common/yang_parse_tree.h"
+#include "stratum/hal/lib/yang/yang_parse_tree.h"
 
 namespace stratum {
 namespace hal {
